requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.12.0
Pillow>=8.3.0
numpy>=1.21.0
matplotlib>=3.4.0
gradio>=2.9.0
fastapi>=0.68.0
uvicorn>=0.15.0
pyyaml>=5.4.0